+++ lib/warnings/parser/bandit_parser.rb
@@ -1,5 +1,5 @@
 require_relative 'parser'
-require_relative '../issue'
+require_relative '../report/issue'
 
 module Warnings
   # Parser class for bandit generated json files.
@@ -33,7 +33,7 @@ module Warnings
       issue.severity = to_severity(hash['issue_severity'])
       issue.message = hash['issue_text']
       issue.line = hash['line_number']
-      issue.id = hash['test_id']
+      issue.category = hash['test_id']
       issue.name = hash['test_name']
       @issues << issue
     end