+++ lib/warnings/report/reporter.rb
@@ -1,6 +1,6 @@
-require_relative 'parser/parser_factory'
-require_relative 'markdown_util'
-require_relative 'severity'
+require_relative '../parser/parser_factory'
+require_relative '../helper/message_util'
+require_relative '../helper/severity'
 
 module Warnings
   # Base reporter class to define attributes and common method to create a report.
@@ -124,7 +124,7 @@ module Warnings
 
     def inline_comment
       @issues.each do |issue|
-        text = inline_text(issue)
+        text = MessageUtil.inline(issue)
         if fail_error && high_issue?(issue)
           @danger.fail(text, line: issue.line, file: issue.file_name)
         else
@@ -133,12 +133,8 @@ module Warnings
       end
     end
 
-    def inline_text(issue)
-      "#{issue.severity.to_s.capitalize}\n[#{issue.id}-#{issue.name}]\n#{issue.message}"
-    end
-
     def markdown_comment
-      text = MarkdownUtil.generate(name, @issues)
+      text = MessageUtil.markdown(name, @issues)
       @danger.markdown(text)
       @danger.fail(format(ERROR_HIGH_SEVERITY, name)) if fail_error && high_issues?
     end