+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added 
+- PylintParser - Support pylint formatted reports
+
+### Changed
+- Renamed issue.id to issue.category
 
 ## [0.0.1] - 2019-01-21
 ### Added